Feature: Home leaflet map and encounter cards
The design of the cards is complete. I like how it looks, I'll get
some feedback later. I want to add some basic UX like onClick and
onHover events.

The basic UX of the map is complete. It loads the basemap, draws
circle markers for each animal from the ALL_ENCOUNTERS_QUERY. I'll
need to add some methods to deal with loading a ton of data, and
keeping it in sync with the encounter cards.

Ref: #10, #11, kissmygritts/gus-graphql/#10