Feature: hover & leave events on encounter cards
This feature adds hover and leave events on the encounter cards.
When hovering over a card the corresponding marker on the map will
be highlighted by a green circle. When the pointer leaves the card
the highlight will be removed. This closes #30.

I have hacked this together. Instead of changing the styling on the
corresponding marker I am simply putting another marker over it. I
cannot think of the proper way to change the style of the marker. If
I had to guess I would attempt to find the marker in the
this.$refs.map.mapObject object. I don't feel like fiddling with
this now. I can add this later as a refactor/code cleanup.

The reactivity added here will prove useful when I start working
on issue #10.

Close: #30
Ref: #10