Feature: Species list page, WIP
I haven't finished and added the level of detail I would like to
have included for the species pages. The UX can be improved: filters,
sorting, etc. A species group page similar to what is seen in CodePen.
I'm going to close #16 for now as it is good enough for v0.2.0 demo
on Tuesday.

Closes: #16